pyproject.toml
[project]
name = "deval"
version = "0.1.0"
description = "Narrow, survivable Polymarket whale-copy bot"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.18.0",
    "polars>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
